Split placeholder refresh tests; extract shared postPaymentLeg helper; add e2e Square refund
Split test/features/admin/attendees-edit.test.ts (410→297 lines) by extracting the placeholder refresh tests into test/features/admin/placeholder-refresh.test.ts (115 lines). Both files now stay under the ~400-line target. Extracted the shared postPaymentLeg helper into test/test-utils/db-helpers/payment-leg.ts.

E2e: the nightly Square sandbox run now exercises the real refund path (afterPaidBooking → exerciseAdminRefund) via the shared shared.ts helper, so the POST /v2/refunds → Valibot schema → COMPLETED check → payment_id/amount verification → ledger posting round-trip is exercised against the real Square sandbox API every night. Extracted exerciseAdminRefund into shared.ts so Stripe and Square share one implementation.
